+++ R/shapeConstr.ps.R
@@ -6,8 +6,8 @@
 ################################################################################
 
 #' @export
-cons_shape.ps <- function(x, diff = 0, sign = 1, ...){
+shapeConstr.ps <- function(x, shape, ...){
 
   # Same as B-Splines
-  cons_shape.bs(x, diff, sign)
+  shapeConstr.bs(x, shape, ...)
 }