+++ man/examples/checkCmat_ex.R
@@ -8,10 +8,10 @@ cmatic <- rbind(diff(diag(p)), diff(diag(p), diff = 2))
 # Checking indicates that constraints 2 to 4 are redundant.
 # Intuitively, if the first two coefficients increase,
 # then convexity forces the rest to increase
-check_cmat(cmatic)
+checkCmat(cmatic)
 
 # Check without contraints
-check_cmat(cmatic[-(2:4),])
+checkCmat(cmatic[-(2:4),])
 
 ###################################################
 # Example of irreducible matrix
@@ -30,7 +30,7 @@ qr(t(cmats))$rank
 all.equal(cmats[2,] + cmats[4,] - cmats[5,], cmats[3,])
 
 # However, it is irreducible: all constraints are necessary
-check_cmat(cmats)
+checkCmat(cmats)
 
 ###################################################
 # Example of underlying equality constraint
@@ -39,4 +39,4 @@ check_cmat(cmats)
 cmateq <- rbind(rep(1, 3), rep(-1, 3))
 
 # Checking indicates that both constraints imply equality constraint (sum == 0)
-check_cmat(cmateq)
+checkCmat(cmateq)