+++ tests/testthat/test-checkCmat.R
@@ -8,15 +8,15 @@
 
 # Three parameters sum to zero
 cmat1 <- rbind(rep(1, 3), rep(-1, 3))
-check1 <- check_cmat(cmat1)
+check1 <- checkCmat(cmat1)
 
 # Another constraint: both parameters non-negative and their sum nonpositive
 # Indicates both are contrained to zero
 cmat2 <- rbind(diag(2), -1)
-check2 <- check_cmat(cmat2)
+check2 <- checkCmat(cmat2)
 
 # Check it is detected
-test_that("check_cmat detects equality constraints", {
+test_that("checkCmat detects equality constraints", {
   expect_equal(check1$equality, c(1, 2))
   expect_equal(check2$equality, 1:3)
 })
@@ -28,7 +28,7 @@ p <- 10
 # Increasing convex
 # because of convexity, the positive difference between coef 2 and 3 on is redundant
 cmat3 <- rbind(diff(diag(p)), diff(diag(p), diff = 2))
-check3 <- check_cmat(cmat3)
+check3 <- checkCmat(cmat3)
 
 # S shape: although not of full row rank, it is irreducible
 cmat4 <- rbind(
@@ -37,10 +37,10 @@ cmat4 <- rbind(
   diff(diag(p), diff = 2)[1:(p/2 - 1),], # First half convex
   -diff(diag(p), diff = 2)[(p/2):(p-2),] # second half concave
 )
-check4 <- check_cmat(cmat4)
+check4 <- checkCmat(cmat4)
 
 # Test
-test_that("check_cmat takes good decisions on irreducibility", {
+test_that("checkCmat takes good decisions on irreducibility", {
   expect_gt(length(check3$redundant), 0)
   expect_length(check4$redundant, 0)
 })